use the bail out interface
